feat(songs): rethink the `MusicPlayer` model
We're moving toward a decentralized structure.
The `Songs` singleton is responsible for creating `SongPlayer`s that in turn handle playback.
In the future, we can potentially use providers to have multiple `SongPlayer`s in the tree, but at the moment the `Songs` singleton makes sure there is only one active player at a time.